refactor(web): rename app folder to application to avoid conflict with Next.js app router
