+++ examples/next-starter/apps/web/src/application/application.ts
