+++ examples/next-starter/apps/web/src/pages/_app.tsx
@@ -4,7 +4,7 @@ if (process.env.NODE_ENV === 'development' && process.env.NEXT_PUBLIC_ENABLE_MIR
 
   makeServer({ environment: 'development', mocks });
 }
-import { application } from '../app/application';
+import { application } from '../application/application';
 import '../assets/styles.css';
 
 const AppComponent = application.getAppComponent();
